user input validation (new strategy) + clean up

+++ app/src/main/kotlin/treasureisland/model/Pirate.kt
@@ -3,7 +3,7 @@ package treasureisland.model
 
 class Pirate(
     val coordinates: Pair<Int, Int>?,
-    val percentOffTreasureVal: Float
+    val percentOffTreasureVal: Int
 ) {
